+++ src/kafka.rs
@@ -1,3 +1,4 @@
+mod conversion;
 mod incremental;
 
 pub use incremental::*;
