+++ src/mongodb/incremental_sync.rs
@@ -386,7 +386,7 @@ pub async fn run_incremental_sync(
             Ok(change) => {
                 debug!("Received change: {change:?}");
 
-                crate::apply_change(&surreal, &change).await?;
+                crate::surreal::apply_change(&surreal, &change).await?;
 
                 change_count += 1;
                 if change_count % 100 == 0 {
