+++ src/neo4j/full_sync.rs
@@ -223,7 +223,7 @@ async fn migrate_neo4j_nodes(
                     label
                 );
                 if !to_opts.dry_run {
-                    crate::write_records(surreal, &label.to_lowercase(), &batch).await?;
+                    crate::surreal::write_records(surreal, &label.to_lowercase(), &batch).await?;
                 } else {
                     tracing::debug!("Dry-run mode: skipping actual migration of batch");
                 }
@@ -242,7 +242,7 @@ async fn migrate_neo4j_nodes(
                 label
             );
             if !to_opts.dry_run {
-                crate::write_records(surreal, &label.to_lowercase(), &batch).await?;
+                crate::surreal::write_records(surreal, &label.to_lowercase(), &batch).await?;
             } else {
                 tracing::debug!("Dry-run mode: skipping actual migration of final batch");
             }
@@ -326,7 +326,8 @@ async fn migrate_neo4j_relationships(
                     rel_type
                 );
                 if !to_opts.dry_run {
-                    crate::write_relations(surreal, &rel_type.to_lowercase(), &batch).await?;
+                    crate::surreal::write_relations(surreal, &rel_type.to_lowercase(), &batch)
+                        .await?;
                 } else {
                     tracing::debug!("Dry-run mode: skipping actual migration of batch");
                 }
@@ -349,7 +350,7 @@ async fn migrate_neo4j_relationships(
                 rel_type
             );
             if !to_opts.dry_run {
-                crate::write_relations(surreal, &rel_type.to_lowercase(), &batch).await?;
+                crate::surreal::write_relations(surreal, &rel_type.to_lowercase(), &batch).await?;
             } else {
                 tracing::debug!("Dry-run mode: skipping actual migration of final batch");
             }
